Fix linking issue #2510
